Changes Db storage

+++ src/main/java/ru/netology/diplom/service/ServiceCloud.java
@@ -1,19 +1,21 @@
 package ru.netology.diplom.service;
 
 import lombok.extern.slf4j.Slf4j;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.core.io.ByteArrayResource;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
-import ru.netology.diplom.Token.JWTUtil;
 import ru.netology.diplom.dto.FileResponse;
+import ru.netology.diplom.exceptions.ErrorFileException;
 import ru.netology.diplom.exceptions.ErrorInputDataException;
 import ru.netology.diplom.model.FileData;
 import ru.netology.diplom.model.RenameFile;
 import ru.netology.diplom.model.UserData;
 import ru.netology.diplom.repository.FileRepository;
 import ru.netology.diplom.repository.UserRepository;
+import ru.netology.diplom.repository.UserTokenRepository;
 
-import java.io.*;
+import java.io.IOException;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.concurrent.atomic.AtomicInteger;
@@ -21,21 +23,17 @@ import java.util.concurrent.atomic.AtomicInteger;
 @Service
 @Slf4j
 public class ServiceCloud {
-    private final FileRepository fileRepository;
-    private final UserRepository userRepository;
-    private final JWTUtil jwtUtil;
-    private String userName;
+    @Autowired
+    private FileRepository fileRepository;
+    @Autowired
+    private UserRepository userRepository;
+    @Autowired
+    private UserTokenRepository tokenRepository;
     private AtomicInteger count = new AtomicInteger(0);
 
-    public ServiceCloud(FileRepository fileRepository, UserRepository userRepository, JWTUtil jwtUtil) {
-        this.fileRepository = fileRepository;
-        this.userRepository = userRepository;
-        this.jwtUtil = jwtUtil;
-    }
-
-    public void fileUpload(String token, String name, MultipartFile file) {
-        userName = jwtUtil.getUsername(jwtUtil.resolveToken(token));
-        UserData user = userRepository.findByEmail(userName);
+    public String fileUpload(String token, String name, MultipartFile file) {
+        UserData user = getUserByAuthToken(token);
+        assert user != null;
         FileData fileDataFromBd = fileRepository.findByFileNameAndUserDataEmail(name, user.getEmail());
         FileData fileData;
         try {
@@ -50,42 +48,65 @@ public class ServiceCloud {
                 fileData.setFileName("(" + count.incrementAndGet() + ")" + name);
             }
             fileRepository.save(fileData);
+            log.info("Файл {} успешно загружен", fileData.getFileName());
+
+            return "Файл успешно загружен";
         } catch (IOException e) {
-            throw new RuntimeException(e);
+            throw new ErrorFileException("Не удалось загрузить файл");
         }
     }
 
     public List<FileResponse> allFiles(String token, Integer limit) {
-        String name = jwtUtil.resolveToken(token);
-        if (name == null) {
+        UserData data = getUserByAuthToken(token);
+        if (data == null) {
             log.error("Ошибка вывода списка всех файлов");
             throw new ErrorInputDataException("Ошибка вывода списка всех файлов");
         }
-        UserData data = userRepository.findByEmail(jwtUtil.getUsername(name));
         log.info("Список файлов успешно выведен на экран");
         return fileRepository.findFileDataByUserDataEmail(data.getEmail()).stream()
                 .map(a -> new FileResponse(a.getFileName(), a.getSize())).toList();
     }
 
     public ByteArrayResource fileDownload(String name, String token) {
-        userName = jwtUtil.getUsername(jwtUtil.resolveToken(token));
-        FileData fileData = fileRepository.findByFileNameAndUserDataEmail(name, userName);
+        UserData userName = getUserByAuthToken(token);
+        if (userName == null) {
+            log.error("Ошибка скачивания файла");
+            throw new ErrorFileException("Ошибка скачивания файла");
+        }
+        FileData fileData = fileRepository.findByFileNameAndUserDataEmail(name, userName.getEmail());
+        log.info("Файл {} успешно скачан", fileData.getFileName());
         return new ByteArrayResource(fileData.getFileData());
     }
 
     public void fileDelete(String fileName, String token) {
-        userName = jwtUtil.getUsername(jwtUtil.resolveToken(token));
-        FileData fileData = fileRepository.findByFileNameAndUserDataEmail(fileName, userName);
+        UserData userName = getUserByAuthToken(token);
+        if (userName == null) {
+            log.error("Не удалось удалить файл");
+            throw new ErrorFileException("Ошибка удаления файла");
+        }
+        FileData fileData = fileRepository.findByFileNameAndUserDataEmail(fileName, userName.getEmail());
         fileRepository.delete(fileData);
         log.info("Файл {} успешно удален", fileData.getFileName());
     }
 
     public void renameFile(String fileName, RenameFile renameFile, String token) {
-        userName = jwtUtil.getUsername(jwtUtil.resolveToken(token));
-        FileData fileData = fileRepository.findByFileNameAndUserDataEmail(fileName, userName);
+        UserData userName = getUserByAuthToken(token);
+        if (userName == null) {
+            log.error("Ошибка переименования файла");
+            throw new ErrorFileException("Ошибка переименования файла");
+        }
+        FileData fileData = fileRepository.findByFileNameAndUserDataEmail(fileName, userName.getEmail());
         fileData.setFileName(renameFile.getFilename());
         fileRepository.save(fileData);
         log.info("Файл {} успешно переименован в {} ", fileName, renameFile.getFilename());
+    }
 
+    private UserData getUserByAuthToken(String authToken) {
+        if (authToken.startsWith("Bearer ")) {
+            final String authTokenWithoutBearer = authToken.split(" ")[1];
+            final String username = tokenRepository.getUserNameByToken(authTokenWithoutBearer);
+            return userRepository.findByEmail(username);
+        }
+        return null;
     }
 }
